Added Player Disconnection Logic
Whenever a player closes the windows, the disconnect event is called in the backend and that player will be removed from the sessions object and will be broadcasted to every remaining player in the room and modifies the local storage and states of every player. Also if the Admin leaves the lobby, admin privileges will be given to the player next to the admin.